[0.6.6] feat(Infrastructure) Added all server-side security on endpoints + added User album attributes endpoints

+++ Musicx.Presentation.Web/Musicx.Presentation.Web/Controllers/AlbumController.cs
@@ -241,7 +241,7 @@ public sealed class AlbumController(IAlbumRepository albumRepository,
     }
 
     [HttpGet("count")]
-    public async Task<ActionResult<int>> GetCount()
+    public async Task<ActionResult<long>> GetCount()
     {
         _logger.LogInformation($"🌍🏳️ API : COUNT albums");
